requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.4
djangorestframework==3.12.2
importlib-metadata==3.3.0
jsonschema==3.2.0
psycopg2==2.8.6
pyrsistent==0.17.3
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
